requirements.txt
streamlit==1.20.0
pandas==1.5.2
mysql-connector-python==8.0.33
requests==2.28.2
python-dotenv==0.21.0